Ensure LayoutInflaterCompat forces set the correct Factory
This is a follow up to Ib9cd690725b365ad40d4a6cf12c1e6dc6c0089fd
which makes sure that any subsequent merged factory is set to be
the designated Factory2. This ensures that we use the merged
factory over another Factory2.

Also bumped up the version number in preparation.

BUG: 20477817

Change-Id: Ie4ea8595afffe9905bb11174f330db4dedb2d2fb