AppCompatWidgets, now with added pixie dust.
This CL moves a lot of the internal code for apply background
tints to a central helper class. This allows us to have one place
to add/fix things. This has enabled an easier implementation of
reading the backgroundTint and backgroundTintMode attributes.

With that, our CompoundButton implementations have followed in
the same vain with a new helper class, reading of the new
buttonTint and buttonTintMode attributes, with a v4 shim to allow
dynamic setting.

Also tidied up the javadoc on each class too so that developers
know what they're getting from each class, and fixed an NPE in
the AutoCompleteTextView classes.

BUG: 19037149
BUG: 21102590

Change-Id: Ice097ed44e1253e5799bbe9ab30285912617b4dc