Not filling accessibility info for hidden views.
This change is distinct from totally removing views from accessibility
radar by marking them as not important for accessibility. That
CR is in progress, and will continue.

This CL just plugs the possibility of a crash when an item is hidden.

It makes sense to land it ASAP to at least plug the worst problem first.

Bug: 20055027
Change-Id: I78ddd16b1cf384d22558255986a4eab34b44269f