NavigationView handles checked status
NavigationView now handles checked status in its menu. This was
necessary to handle checked status for items in sub menu. Also, it now
forces only one item to be marked as checked among all the checkable
items in the menu. Menu groups can be marked as checkedBehavior="single"
or ="all", but all the items behave the same in either cases.

Change-Id: Ia2e1a39475175b5ada8ec84273a39b995ec414c3