Fix options list menu when Toolbar doesn't have items
When using a ToolbarActionBar, and there are 0 non-actions
items, it will return null in onCreatePanelView. This then
means that the 'window' will show it's list menu. As the
'window' doesn't know what is an action item or not,
it shows all items.

This CL fixes it by disabling the window list menu view
when a ToolbarActionBar is used. ToolbarActionBar's own
list menu view is unaffected.

BUG: 21718389
Change-Id: I8e8cb7915cc27abcf1e06a0ef09ac14129c4adce