Fix LinearSmoothScroller NPE bug
This CL fixes a bug where LinearSmoothScroller would try to access
RecyclerView after it is stopped.

This was actually caused by a bad API so I deprecated it in favor
of a more open API. This new API, jumpTo also allows instantly
scrolling to a position in case target position is far away.

Bug: 21661478
Change-Id: I3c4bfa8589803a564a61116e73d8b1312c6c9f46