DesignTime: Do not use animation in design time.
Animations are used in response to a nested scroll in an App Bar
to: turn off the FAB and make the scrim opaque such that a image
is invisible when the toolbar is collapsed.
In design time: do not use animations.
This will allow LayoutLib to render the controls correctly after
the toolbar is collapsed.

Change-Id: Id168ce24d4539a765ccbe09a11ffc96716f28ae6