Unregister ItemBridgeAdapter from ObjectAdapter
This is a rare use case: if same ObjectAdapter is reused
across multiple Fragments and backstack is not supported.
(Activity switch fragment showing same data?)
We must remove ItemBridgeAdapter from observer list of
the long existing ObjectAdapter.

Fragment should in theory to be self restoreable for example
when BACK is supported. So ObjectAdapter should have same
lifecycle as RowsFragment.  This fix isnt necessary in
these typical cases.

But it is a safe fix.

Bug: 22642503

Change-Id: Id0b40a0c1b022d0ec534f6281fb38156a1436c06