MediaSessionCompat: fix the flaky remote volume slider issue
RemoteControlClient keeps the state locally and send an update to AudioService
only when the state changes. The problem is that the state in the AudioService
is reset when we register RemoteControlClient while RemoteControlClient keeps
the state. Discrepancy happens when we re-register the RemoteControlClient
instance. This change resets the state in RemoteControlClient so that the state
in AudioService has the right value when we re-register.

Bug: 19719195
Change-Id: Ic3b714919be35ef1fa6f1dc698c8811387db787d