customtabs: Add "extraCommand" and "extraCallback", for unsupported commands
This adds a new AIDL method, extraCommand(String, Bundle). It is
intended to allow implementations to add extra unsupported commands
forcetesting or experimental purposes. These are not intended to be
documented, and should be either internal (for testing) or transient (as
a testing ground before adding them as a proper AIDL method).

In Chromium's case, this command will be used for these two purposes.

In the same fashion, this also introduces an "extra" callback.

Change-Id: I468a6f7bbf82a83bb480ff41e1239dc8c8254a40