Allow NestedScrollView to always intercept touches
Currently if a NestedScrollView's content aren't long
enough to scroll then it doesn't intercept any touch
events. That is fine for it's own scrolling but can
interfere with things when nested scrolling is in the
mix.

By not intercepting any touch events, nested scrolling
will not happen if the touch event originate from a child
view which has it's own touch handling (i.e. a clickable
Button).

BUG: 23380443
Change-Id: Ieec61ecb193068a538623a8c5fbdb08efea2c6a5