Fix ABL + CTL not handling being initially collapsed well
Two issues here. The first being that ABL only updates it's
min height after a layout, which means it's minHeight is wrong
in the preceding onMeasure(). Fixed by invalidating the scroll
ranges in onLayout, and updating it's minHeight in onMeasure.

Second issue is that CTL's text helper needs it's bounds in
the initial coordinate space, not after the views have been
offset.

Also fixed a possible pending state issue.

BUG: 23408061
Change-Id: Ie15c2e353da1596ca045405a6cf6f9cd2209a32e