MediaRouter: Update the art view visibility only when it's necessary
Recent changes made around the art view introduced regression where the
entire controller dialog shakes too much as we remove the art view and
then immediately add it again every time there is update on the media
metadata or playback state. This change ensures we hide or unhide the
art view only when it's needed.

Bug: 23505321, Bug: 23758462
Change-Id: I8815186bfd9c98113fbab7701916e44e19c18c24