MediaRouter: expand group volume section as much as it can
Previouly, when the artwork is 16:9, the group volume section had scroller
even when it fully covers the artwork. This change expands the section as much
as it can to remove the scrolling.

Also, onGlobalLayout mechanism is removed and decor view measuring approach is
used instead since the former one doesn't work well with the built-in dialog
animation.

Bug: 23118305, Bug: 23117755
Change-Id: Ibe1c03b6103f45fb90917661fb53fc57fde49c3e