Update NavigationView to use CoordinatorLayout for header and list.
This CL modifies how the header is handled in NavigationView. Namely, it
separates the header from the ListView. To scroll the header and list
together, both the header and list are placed in a CoordinatorLayout with
behaviors that allow them to scroll together.

This work will allow an overlay to be added that scrolls with the list
so that we can properly animate the account switcher. The overlay and
account switcher work will be done in a separate CL.

Bug: 20117594
Change-Id: Ic72c86f165e26c493b965dedf4c65f2853409144