MediaRouter: Implement the volume group animaion without window animation
Note: when animating the dialog size, we need to fix the window size since
the animation will be jerky without it because view animation and window animation
do not happen in sync. To address this, this change creates a transparent layout
which expands the maximum area of possible animation and animate the dialog
within the area.

Bug: 24271568
Change-Id: I1e691cbf3124205f55624a4a04d713206d6be528