Fix content shifting when an Action Mode is displayed
Caused by a small timing issue in the new fade animations.
The context view is GONE initially so may not have been
laid out when the animation actually starts. This can lead
to the container collapsing to 0px height for a short period.

Fixed by using INVISIBLE for the Toolbar instead to maintain
the height.

BUG: 24395058

Change-Id: I3ec3b0c04c6cd2258692efb2c3d879cec6a82875