Cleanup for item tracking
This CL abstracts the Item tracking for RecyclerView. This makes
it much easier to test (better encapsulation) + cleaner code.
And as an added bonus, I merged some maps to speedup lookups.

There is also a small (unreleased) API change for the new
ItemAnimation callbacks. It was wrong in the first one but didn't
realize because SimpleItemAnimator was taking care of it.

I've also added some support annotations & fixed some warnings.

Bug: 24665726
Change-Id: Idba09e749f1c0f269d81dbdee4056f6b52ba8484