Add AppCompatDrawableManager
Refactored from TintManager, AppCompatDrawableManager
is now architectured differently. It is now a singleton
which keeps multi-map of cached tints.

This is to enable the new InflateDelegate extension point,
which will allows libraries (and eventually apps) to hook
into the compat drawable inflation.

While this CL looks large, it's mostly just updating method
calls to match the new parameters.

BUG: 24852773

Change-Id: Ic9f5ecc6fd0208d2f7a8a686f077f16588020516