Allow swiped view recycling ASAP
This CL fixes a bug in ItemTouchHelper where it would assume that the
swiped view will be removed by the RecyclerView even if its animation
count does not reach 0.

This was a bad assumption and broke w/ the RecyclerView cleanups.

This CL changes ItemTouchHelper to allow view's recycling instantly
but do the cleanup when the view is detached (e.g. all other
animations etc. allows it to be removed).

This CL also fixes a bug in ItemAnimatorV2 test where it would call dispatch
end twice for change animations.

Bug: 25037922
Change-Id: I8c264ef1c7a89cb03f967796a0e14739f9e1945b