GuidedStepFragment: fix initial selector view size flicking
In layout pass, onLayout of VerticalGridView triggers a focus change event
where we change LayoutParams of the selectorView, but because parent RelativeLayout
does not re-measure the selectorView in RelativeLayout.onLayout, parent RelativeLayout
calls selectorView.layout() with the obsolete values and leave the view size
inconsistent with LayoutParams.

Our current workaround "post a Runnable" causes one frame flicking when fragment
is initially loaded.

One non-trivial fix is to override RelativeLayout and re-measure the selectorView
between VerticalGridView.onLayout() and selectorView.onLayout().

But since we use scaleY anyway later,  switching to use scaleY is much simpler.

Change-Id: I585609695048cffd23e67619320af187e37ffbb6