MediaSessionCompat: Protect artwork bitmaps from being recycled
When an artwork image is updated on pre-L, current RCC implementation
recycles an original one, making it inaccessible later. In order to make
this behavior fit well with caching, this change clones given bitmaps in
the metadata to be set. An original artwork bitmap can remain
unrecycled, and RCC will recycle the cloned one later.

Bug: 24540321
Change-Id: I7f993330252d55d55884b6195a1f79e9873bfa62