Fix NSV not allowing clicks after a small fling
This was a fun one to work out. The reason that manifests
is because NSV's onInterceptTouchEvent checks if it's
scroller.isFinished() returns true.

NSV was missing some code to manually post an invalidation,
and thus also a further call to computeScroll(). This means
that the result of Scroller.isFinished() is incorrect since
it hasn't had computeScrollOffset() called on it since the
last scroll event (which could be quite a while ago).

This CL fixes this by adding the invalidation post, but also
adds a call to computeScrollOffset() before calling isFinished().

BUG: 26051101
Change-Id: Iad15a165f6c8605bcd98efe2869df955792c4f1b