Try to detect developer error for change animations
If an adapter does not have proper stable ids, it may crash RV while
it tries to animate. This is NOT a RV bug but it is very hard for the
developer to recognize this since the crash does not give any clue on
what is wrong.

This CL detects this case and prints a specific error. It also changes
how RV handles this unexpected case. Now, if it cannot detect the error,
it will just log instead of crashing. It will still crash if it detects
an error in the Adapter.

Bug: 25715473
Change-Id: I896bceb09417a42899951311352c15f7c912c57b