GuidedStepFragment: refactoring the focusable control
1. The actions list view does not need to be focusable.

2. It's a pain to disable focus left/right at verticalgridview
level.  For example when button actions are not visible, we
should disable "focus out of right edge of main actions",  when
button actions are visible, we should enable "focus out of right
edge of main actions". Moving the disable left/right logic to
top layout make it lot cleaner.

3. Allow GuidedStepFragment subclass to customize isFocusOutStartAllowed
or isFocusOutEndAllowed

Bug: 26205201

Change-Id: I66ee6bee8c7abbfe97db0f9fdf8119615143f543